# docs/swagger.yaml
definitions:
...
  apiLink.FriendAddRequest:
    properties:
      link_apply_remark:
        example: 申请友链
        maxLength: 500
        type: string
      link_avatar:
        example: https://example.com/avatar.jpg
        maxLength: 500
        type: string
      link_color_id:
        example: 1
        type: integer
      link_desc:
        example: 这是一个示例网站
        maxLength: 500
        type: string
      link_email:
        example: admin@example.com
        maxLength: 100
        type: string
      link_group_id:
        example: 1
        type: integer
      link_name:
        example: 示例网站
        maxLength: 100
        minLength: 1
        type: string
      link_order:
        example: 0
        minimum: 0
        type: integer
      link_rss:
        example: https://example.com/rss.xml
        maxLength: 500
        type: string
      link_url:
        example: https://example.com
        maxLength: 500
        type: string
    required:
    - link_name
    - link_url
    type: object
  apiLink.FriendAddResponse:
    properties:
      apply_remark:
        description: 申请者备注
        type: string
      avatar:
        description: 友链头像URL
        type: string
      color_f_key:
        allOf:
        - $ref: '#/definitions/entity.LinkColor'
        description: 友链颜色外键
      color_id:
        description: 友链颜色ID
        type: integer
      created_at:
        description: 创建时间
        type: string
      description:
        description: 友链描述
        type: string
      email:
        description: 友链联系邮箱
        type: string
      fail_reason:
        description: 友链失效原因
        type: string
      group_f_key:
        allOf:
        - $ref: '#/definitions/entity.LinkGroup'
        description: 关联关系
      group_id:
        description: 所属分组ID
        type: integer
      id:
        description: ID 使用 Snowflake ID 作为主键
        type: integer
      is_failure:
        description: 友链失效标志
        type: integer
      name:
        description: 友链名称
        type: string
      review_remark:
        description: 审核备注
        type: string
      rss:
        description: 友链RSS地址
        type: string
      sort_order:
        description: 友链排序
        type: integer
      status:
        description: 友链状态
        type: integer
      updated_at:
        description: 更新时间
        type: string
      url:
        description: 友链URL地址
        type: string
    type: object
  apiLink.FriendDetailResponse:
    properties:
      apply_remark:
        description: 申请者备注
        type: string
      avatar:
        description: 友链头像URL
        type: string
      color_f_key:
        allOf:
        - $ref: '#/definitions/entity.LinkColor'
        description: 友链颜色外键
      color_id:
        description: 友链颜色ID
        type: integer
      created_at:
        description: 创建时间
        type: string
      description:
        description: 友链描述
        type: string
      email:
        description: 友链联系邮箱
        type: string
      fail_reason:
        description: 友链失效原因
        type: string
      group_f_key:
        allOf:
        - $ref: '#/definitions/entity.LinkGroup'
        description: 关联关系
      group_id:
        description: 所属分组ID
        type: integer
      id:
        description: ID 使用 Snowflake ID 作为主键
        type: integer
      is_failure:
        description: 友链失效标志
        type: integer
      name:
        description: 友链名称
        type: string
      review_remark:
        description: 审核备注
        type: string
      rss:
        description: 友链RSS地址
        type: string
      sort_order:
        description: 友链排序
        type: integer
      status:
        description: 友链状态
        type: integer
      updated_at:
        description: 更新时间
        type: string
      url:
        description: 友链URL地址
        type: string
    type: object
  apiLink.FriendFailRequest:
    properties:
      link_fail:
        enum:
        - 0
        - 1
        example: 1
        type: integer
      link_fail_reason:
        example: 网站无法访问
        maxLength: 500
        type: string
    required:
    - link_fail
    type: object
  apiLink.FriendListResponse:
    properties:
      data:
        description: 数据列表
        items:
          $ref: '#/definitions/entity.LinkFriend'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/base.PaginationInfo'
        description: 分页信息
    type: object
  apiLink.FriendPublicResponse:
    properties:
      links:
        items:
          $ref: '#/definitions/entity.LinkFriend'
        type: array
    type: object
  apiLink.FriendStatusRequest:
    properties:
      link_review_remark:
        example: 审核通过
        maxLength: 500
        type: string
      link_status:
        enum:
        - 0
        - 1
        - 2
        example: 1
        type: integer
    required:
    - link_status
    type: object
  apiLink.FriendUpdateRequest:
    properties:
      link_apply_remark:
        example: 申请友链
        maxLength: 500
        type: string
      link_avatar:
        example: https://example.com/avatar.jpg
        maxLength: 500
        type: string
      link_color_id:
        example: 1
        type: integer
      link_desc:
        example: 这是一个示例网站
        maxLength: 500
        type: string
      link_email:
        example: admin@example.com
        maxLength: 100
        type: string
      link_group_id:
        example: 1
        type: integer
      link_name:
        example: 示例网站
        maxLength: 100
        minLength: 1
        type: string
      link_order:
        example: 0
        minimum: 0
        type: integer
      link_rss:
        example: https://example.com/rss.xml
        maxLength: 500
        type: string
      link_url:
        example: https://example.com
        maxLength: 500
        type: string
    type: object
  apiLink.FriendUpdateResponse:
    properties:
      apply_remark:
        description: 申请者备注
        type: string
      avatar:
        description: 友链头像URL
        type: string
      color_f_key:
        allOf:
        - $ref: '#/definitions/entity.LinkColor'
        description: 友链颜色外键
      color_id:
        description: 友链颜色ID
        type: integer
      created_at:
        description: 创建时间
        type: string
      description:
        description: 友链描述
        type: string
      email:
        description: 友链联系邮箱
        type: string
      fail_reason:
        description: 友链失效原因
        type: string
      group_f_key:
        allOf:
        - $ref: '#/definitions/entity.LinkGroup'
        description: 关联关系
      group_id:
        description: 所属分组ID
        type: integer
      id:
        description: ID 使用 Snowflake ID 作为主键
        type: integer
      is_failure:
        description: 友链失效标志
        type: integer
      name:
        description: 友链名称
        type: string
      review_remark:
        description: 审核备注
        type: string
      rss:
        description: 友链RSS地址
        type: string
      sort_order:
        description: 友链排序
        type: integer
      status:
        description: 友链状态
        type: integer
      updated_at:
        description: 更新时间
        type: string
      url:
        description: 友链URL地址
        type: string
    type: object
...
  apiLink.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  base.PaginationInfo:
    properties:
      has_next:
        description: 是否有下一页
        type: boolean
      has_prev:
        description: 是否有上一页
        type: boolean
      page:
        description: 当前页码
        type: integer
      page_size:
        description: 每页数量
        type: integer
      total:
        description: 总记录数
        type: integer
      total_pages:
        description: 总页数
        type: integer
    type: object
  entity.LinkColor:
    properties:
      created_at:
        description: 创建时间
        type: string
      hover_color:
        description: 悬停颜色
        type: string
      id:
        description: ID 使用 Snowflake ID 作为主键
        type: integer
      links_f_key:
        description: 关联关系
        items:
          $ref: '#/definitions/entity.LinkFriend'
        type: array
      name:
        description: 颜色名称
        type: string
      primary_color:
        description: 主颜色
        type: string
      sort_order:
        description: 颜色排序
        type: integer
      status:
        description: 颜色状态
        type: boolean
      sub_color:
        description: 副颜色
        type: string
      type:
        description: 颜色类型
        type: integer
      updated_at:
        description: 更新时间
        type: string
    type: object
  entity.LinkFriend:
    properties:
      apply_remark:
        description: 申请者备注
        type: string
      avatar:
        description: 友链头像URL
        type: string
      color_f_key:
        allOf:
        - $ref: '#/definitions/entity.LinkColor'
        description: 友链颜色外键
      color_id:
        description: 友链颜色ID
        type: integer
      created_at:
        description: 创建时间
        type: string
      description:
        description: 友链描述
        type: string
      email:
        description: 友链联系邮箱
        type: string
      fail_reason:
        description: 友链失效原因
        type: string
      group_f_key:
        allOf:
        - $ref: '#/definitions/entity.LinkGroup'
        description: 关联关系
      group_id:
        description: 所属分组ID
        type: integer
      id:
        description: ID 使用 Snowflake ID 作为主键
        type: integer
      is_failure:
        description: 友链失效标志
        type: integer
      name:
        description: 友链名称
        type: string
      review_remark:
        description: 审核备注
        type: string
      rss:
        description: 友链RSS地址
        type: string
      sort_order:
        description: 友链排序
        type: integer
      status:
        description: 友链状态
        type: integer
      updated_at:
        description: 更新时间
        type: string
      url:
        description: 友链URL地址
        type: string
    type: object
  entity.LinkGroup:
    properties:
      created_at:
        description: 创建时间
        type: string
      description:
        description: 分组描述
        type: string
      id:
        description: ID 使用 Snowflake ID 作为主键
        type: integer
      links_f_key:
        description: 关联关系
        items:
          $ref: '#/definitions/entity.LinkFriend'
        type: array
      name:
        description: 分组名称
        type: string
      sort_order:
        description: 分组排序
        type: integer
      status:
        description: 分组状态
        type: boolean
      updated_at:
        description: 更新时间
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/admin/links:
    get:
      consumes:
      - application/json
      description: 分页查询友情链接列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 友情链接名称
        in: query
        name: link_name
        type: string
      - description: 友情链接状态 0:待审核 1:已通过 2:已拒绝
        in: query
        name: link_status
        type: integer
      - description: 失效状态 0:正常 1:失效
        in: query
        name: link_fail
        type: integer
      - description: 分组ID
        format: int64
        in: query
        name: link_group_id
        type: integer
      - description: 排序字段
        enum:
        - created_at
        - updated_at
        - link_order
        - link_name
        in: query
        name: sort_by
        type: string
      - description: 排序方式
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/apiLink.FriendListResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 获取友情链接列表
      tags:
      - 友情链接管理
    post:
      consumes:
      - application/json
      description: 添加新的友情链接申请
      parameters:
      - description: 添加友情链接请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apiLink.FriendAddRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 添加成功
          schema:
            $ref: '#/definitions/apiLink.FriendAddResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 添加友情链接
      tags:
      - 友情链接管理
  /api/v1/admin/links/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的友情链接
      parameters:
      - description: 友情链接ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/apiLink.MessageResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 友情链接不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 删除友情链接
      tags:
      - 友情链接管理
    get:
      consumes:
      - application/json
      description: 获取指定友情链接的详细信息
      parameters:
      - description: 友情链接ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/apiLink.FriendDetailResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 友情链接不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 获取友情链接详情
      tags:
      - 友情链接管理
    put:
      consumes:
      - application/json
      description: 更新指定的友情链接信息
      parameters:
      - description: 友情链接ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 更新友情链接请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apiLink.FriendUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/apiLink.FriendUpdateResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 友情链接不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 更新友情链接
      tags:
      - 友情链接管理
  /api/v1/admin/links/{id}/fail:
    put:
      consumes:
      - application/json
      description: 标记友情链接为失效或恢复正常
      parameters:
      - description: 友情链接ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 更新失效状态请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apiLink.FriendFailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/apiLink.MessageResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 友情链接不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 更新友情链接失效状态
      tags:
      - 友情链接管理
  /api/v1/admin/links/{id}/status:
    put:
      consumes:
      - application/json
      description: 审核友情链接，更新状态为通过或拒绝
      parameters:
      - description: 友情链接ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 更新状态请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apiLink.FriendStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/apiLink.MessageResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 友情链接不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 更新友情链接状态
      tags:
      - 友情链接管理
...
  /api/v1/public/links:
    get:
      consumes:
      - application/json
      description: 获取已通过审核且正常的友情链接列表，用于前台展示
      parameters:
      - description: 分组ID，不传则获取所有
        format: int64
        in: query
        name: group_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/apiLink.FriendPublicResponse'
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取公开友情链接
      tags:
      - 公开接口
...
swagger: "2.0"