# internal/handler/link.go
/*
 * --------------------------------------------------------------------------------
 * Copyright (c) 2016-NOW(至今) 筱锋
 * Author: 筱锋「xiao_lfeng」(https://www.x-lf.com)
 * --------------------------------------------------------------------------------
 * 许可证声明：版权所有 (c) 2016-2025 筱锋。保留所有权利。
 * 有关MIT许可证的更多信息，请查看项目根目录下的LICENSE文件或访问：
 * https://opensource.org/licenses/MIT
 * --------------------------------------------------------------------------------
 */

package handler

import (
	apiLink "github.com/bamboo-services/bamboo-main/api/link"

	xValid "github.com/bamboo-services/bamboo-base-go/common/validator"
	xResult "github.com/bamboo-services/bamboo-base-go/major/result"
	"github.com/gin-gonic/gin"
)

// Add 添加友情链接
// @Summary 添加友情链接
// @Description 添加新的友情链接申请
// @Tags 友情链接管理
// @Accept json
// @Produce json
// @Security Bearer
// @Param request body apiLink.FriendAddRequest true "添加友情链接请求"
// @Success 200 {object} apiLink.FriendAddResponse "添加成功"
// @Failure 400 {object} map[string]interface{} "请求参数错误"
// @Failure 401 {object} map[string]interface{} "未认证"
// @Failure 500 {object} map[string]interface{} "服务器内部错误"
// @Router /api/v1/admin/links [post]
func (h *LinkHandler) Add(c *gin.Context) {
	var req apiLink.FriendAddRequest

	// 绑定请求数据
	bindErr := c.ShouldBindJSON(&req)
	if bindErr != nil {
		xValid.HandleValidationError(c, bindErr)
		return
	}

	// 调用服务层
	link, err := h.service.linkLogic.Add(c, &req)
	if err != nil {
		_ = c.Error(err)
		return
	}

	// 返回成功响应
	resp := apiLink.FriendAddResponse{LinkFriend: *link}
	xResult.SuccessHasData(c, "友情链接添加成功", resp)
}

// Update 更新友情链接
// @Summary 更新友情链接
// @Description 更新指定的友情链接信息
// @Tags 友情链接管理
// @Accept json
// @Produce json
// @Security Bearer
// @Param id path int64 true "友情链接ID"
// @Param request body apiLink.FriendUpdateRequest true "更新友情链接请求"
// @Success 200 {object} apiLink.FriendUpdateResponse "更新成功"
// @Failure 400 {object} map[string]interface{} "请求参数错误"
// @Failure 401 {object} map[string]interface{} "未认证"
// @Failure 404 {object} map[string]interface{} "友情链接不存在"
// @Failure 500 {object} map[string]interface{} "服务器内部错误"
// @Router /api/v1/admin/links/{id} [put]
func (h *LinkHandler) Update(c *gin.Context) {
	linkIDStr := c.Param("id")
	var req apiLink.FriendUpdateRequest

	// 绑定请求数据
	bindErr := c.ShouldBindJSON(&req)
	if bindErr != nil {
		xValid.HandleValidationError(c, bindErr)
		return
	}

	// 调用服务层
	link, err := h.service.linkLogic.Update(c, linkIDStr, &req)
	if err != nil {
		_ = c.Error(err)
		return
	}

	// 返回成功响应
	resp := apiLink.FriendUpdateResponse{LinkFriend: *link}
	xResult.SuccessHasData(c, "友情链接更新成功", resp)
}

// Delete 删除友情链接
// @Summary 删除友情链接
// @Description 删除指定的友情链接
// @Tags 友情链接管理
// @Accept json
// @Produce json
// @Security Bearer
// @Param id path int64 true "友情链接ID"
// @Success 200 {object} apiLink.MessageResponse "删除成功"
// @Failure 400 {object} map[string]interface{} "请求参数错误"
// @Failure 401 {object} map[string]interface{} "未认证"
// @Failure 404 {object} map[string]interface{} "友情链接不存在"
// @Failure 500 {object} map[string]interface{} "服务器内部错误"
// @Router /api/v1/admin/links/{id} [delete]
func (h *LinkHandler) Delete(c *gin.Context) {
	linkIDStr := c.Param("id")

	// 调用服务层
	err := h.service.linkLogic.Delete(c, linkIDStr)
	if err != nil {
		_ = c.Error(err)
		return
	}

	// 返回成功响应
	xResult.Success(c, "友情链接删除成功")
}

// Get 获取友情链接详情
// @Summary 获取友情链接详情
// @Description 获取指定友情链接的详细信息
// @Tags 友情链接管理
// @Accept json
// @Produce json
// @Security Bearer
// @Param id path int64 true "友情链接ID"
// @Success 200 {object} apiLink.FriendDetailResponse "获取成功"
// @Failure 400 {object} map[string]interface{} "请求参数错误"
// @Failure 401 {object} map[string]interface{} "未认证"
// @Failure 404 {object} map[string]interface{} "友情链接不存在"
// @Failure 500 {object} map[string]interface{} "服务器内部错误"
// @Router /api/v1/admin/links/{id} [get]
func (h *LinkHandler) Get(c *gin.Context) {
	linkIDStr := c.Param("id")

	// 调用服务层
	link, err := h.service.linkLogic.Get(c, linkIDStr)
	if err != nil {
		_ = c.Error(err)
		return
	}

	// 返回成功响应
	resp := apiLink.FriendDetailResponse{LinkFriend: *link}
	xResult.SuccessHasData(c, "获取成功", resp)
}

// List 获取友情链接列表
// @Summary 获取友情链接列表
// @Description 分页查询友情链接列表
// @Tags 友情链接管理
// @Accept json
// @Produce json
// @Security Bearer
// @Param page query int false "页码" default(1)
// @Param page_size query int false "每页数量" default(10)
// @Param link_name query string false "友情链接名称"
// @Param link_status query int false "友情链接状态 0:待审核 1:已通过 2:已拒绝"
// @Param link_fail query int false "失效状态 0:正常 1:失效"
// @Param link_group_id query int64 false "分组ID"
// @Param sort_by query string false "排序字段" Enums(created_at, updated_at, link_order, link_name)
// @Param sort_order query string false "排序方式" Enums(asc, desc)
// @Success 200 {object} apiLink.FriendListResponse "获取成功"
// @Failure 400 {object} map[string]interface{} "请求参数错误"
// @Failure 401 {object} map[string]interface{} "未认证"
// @Failure 500 {object} map[string]interface{} "服务器内部错误"
// @Router /api/v1/admin/links [get]
func (h *LinkHandler) List(c *gin.Context) {
	var req apiLink.FriendQueryRequest

	// 绑定查询参数
	bindErr := c.ShouldBindQuery(&req)
	if bindErr != nil {
		xValid.HandleValidationError(c, bindErr)
		return
	}

	// 调用服务层
	result, err := h.service.linkLogic.List(c, &req)
	if err != nil {
		_ = c.Error(err)
		return
	}

	// 返回成功响应
	resp := apiLink.FriendListResponse{PaginationResponse: *result}
	xResult.SuccessHasData(c, "获取成功", resp)
}

// UpdateStatus 更新友情链接状态
// @Summary 更新友情链接状态
// @Description 审核友情链接，更新状态为通过或拒绝
// @Tags 友情链接管理
// @Accept json
// @Produce json
// @Security Bearer
// @Param id path int64 true "友情链接ID"
// @Param request body apiLink.FriendStatusRequest true "更新状态请求"
// @Success 200 {object} apiLink.MessageResponse "更新成功"
// @Failure 400 {object} map[string]interface{} "请求参数错误"
// @Failure 401 {object} map[string]interface{} "未认证"
// @Failure 404 {object} map[string]interface{} "友情链接不存在"
// @Failure 500 {object} map[string]interface{} "服务器内部错误"
// @Router /api/v1/admin/links/{id}/status [put]
func (h *LinkHandler) UpdateStatus(c *gin.Context) {
	linkIDStr := c.Param("id")
	var req apiLink.FriendStatusRequest

	// 绑定请求数据
	bindErr := c.ShouldBindJSON(&req)
	if bindErr != nil {
		xValid.HandleValidationError(c, bindErr)
		return
	}

	// 调用服务层
	err := h.service.linkLogic.UpdateStatus(c, linkIDStr, &req)
	if err != nil {
		_ = c.Error(err)
		return
	}

	// 返回成功响应
	xResult.Success(c, "状态更新成功")
}

// UpdateFailStatus 更新友情链接失效状态
// @Summary 更新友情链接失效状态
// @Description 标记友情链接为失效或恢复正常
// @Tags 友情链接管理
// @Accept json
// @Produce json
// @Security Bearer
// @Param id path int64 true "友情链接ID"
// @Param request body apiLink.FriendFailRequest true "更新失效状态请求"
// @Success 200 {object} apiLink.MessageResponse "更新成功"
// @Failure 400 {object} map[string]interface{} "请求参数错误"
// @Failure 401 {object} map[string]interface{} "未认证"
// @Failure 404 {object} map[string]interface{} "友情链接不存在"
// @Failure 500 {object} map[string]interface{} "服务器内部错误"
// @Router /api/v1/admin/links/{id}/fail [put]
func (h *LinkHandler) UpdateFailStatus(c *gin.Context) {
	linkIDStr := c.Param("id")
	var req apiLink.FriendFailRequest

	// 绑定请求数据
	bindErr := c.ShouldBindJSON(&req)
	if bindErr != nil {
		xValid.HandleValidationError(c, bindErr)
		return
	}

	// 调用服务层
	err := h.service.linkLogic.UpdateFailStatus(c, linkIDStr, &req)
	if err != nil {
		_ = c.Error(err)
		return
	}

	// 返回成功响应
	xResult.Success(c, "失效状态更新成功")
}

// GetPublicLinks 获取公开的友情链接
// @Summary 获取公开友情链接
// @Description 获取已通过审核且正常的友情链接列表，用于前台展示
// @Tags 公开接口
// @Accept json
// @Produce json
// @Param group_id query int64 false "分组ID，不传则获取所有"
// @Success 200 {object} apiLink.FriendPublicResponse "获取成功"
// @Failure 500 {object} map[string]interface{} "服务器内部错误"
// @Router /api/v1/public/links [get]
func (h *LinkHandler) GetPublicLinks(c *gin.Context) {
	groupIDStr := c.Query("group_id")

	// 调用服务层
	links, err := h.service.linkLogic.GetPublicLinks(c, groupIDStr)
	if err != nil {
		_ = c.Error(err)
		return
	}

	// 返回成功响应
	resp := apiLink.FriendPublicResponse{Links: links}
	xResult.SuccessHasData(c, "获取成功", resp)
}